Changed getByUuid to returning an optional

+++ src/main/java/com/example/resource/MovieResource.java
@@ -2,6 +2,7 @@ package com.example.resource;
 
 import com.example.dto.MovieDto;
 import com.example.dto.Movies;
+import com.example.entity.Movie;
 import com.example.repository.MovieRepository;
 import jakarta.inject.Inject;
 import jakarta.validation.Valid;
@@ -13,6 +14,7 @@ import jakarta.ws.rs.core.UriInfo;
 
 import java.net.URI;
 import java.time.LocalDateTime;
+import java.util.Optional;
 import java.util.UUID;
 
 @Path("/movies")
@@ -42,15 +44,16 @@ public class MovieResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("{uuid}")
     public MovieDto one(@PathParam("uuid") UUID uuid) {
-        var movie = movieRepository.findByUuid(uuid);
-        if (movie == null)
-            throw new NotFoundException("Invalid id " + uuid);
-        return MovieDto.map(movie);
+        Optional<Movie> m = movieRepository.getByUuid(uuid);
+        if (m.isPresent()) {
+            Movie movie = m.get();
+            return MovieDto.map(movie);
+        } else throw new NotFoundException("No movie found with UUID " + uuid);
     }
 
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response create(@Valid MovieDto movieDto){
+    public Response create(@Valid MovieDto movieDto) {
         var movie = movieRepository.add(MovieDto.map(movieDto));
         return Response.created(URI.create(uriInfo.getAbsolutePath().toString() + "/" + movie.getUuid())).build();
     }
@@ -58,15 +61,14 @@ public class MovieResource {
     @DELETE
     @Path("/{uuid}")
     public Response delete(@PathParam("uuid") UUID uuid) {
-        Response movieDeleteResponse = movieRepository.deleteByUuid(uuid);
-        if (movieDeleteResponse.getStatus() == Response.Status.OK.getStatusCode()) {
-            return Response.ok("Movie successfully deleted").build();
-        } else if (movieDeleteResponse.getStatus() == Response.Status.NOT_FOUND.getStatusCode()) {
+        if (movieRepository.getByUuid(uuid).isPresent()) {
+            movieRepository.deleteByUuid(uuid);
+            return Response.ok("Successfully deleted").build();
+        } else if (!movieRepository.getByUuid(uuid).isPresent()) {
             return Response.status(Response.Status.NOT_FOUND)
                 .entity("No movie found with UUID " + uuid)
                 .build();
-        }
-        return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
+        } else return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
             .entity("Failed to delete movie with UUID: " + uuid)
             .build();
     }
@@ -75,8 +77,17 @@ public class MovieResource {
     @Path("/{uuid}")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response updateOne(@PathParam("uuid") UUID uuid, MovieDto movie) {
-        movieRepository.replace(uuid, MovieDto.map(movie));
-        return Response.created(URI.create("movies/" + uuid)).build();
+        Optional<Movie> m = movieRepository.getByUuid(uuid);
+        if (m.isPresent()) {
+            movieRepository.replace(uuid, MovieDto.map(movie));
+            return Response.ok("Successfully updated").build();
+        } else if (!m.isPresent()) {
+            return Response.status(Response.Status.NOT_FOUND)
+                .entity("No movie found with UUID " + uuid)
+                .build();
+        } else return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
+            .entity("Failed to update movie with UUID: " + uuid)
+            .build();
     }
 
 }
